requirement.txt
Django>=4.2
djangorestframework
django-environ
django-cors-headers
drf-yasg
celery
mysqlclient
django-seed